[expo-notifications] Rename NotificationPresenter to …Module
